requirements.txt
flask
Flask-SQLAlchemy
python-dotenv
flask_jwt_extended
Werkzeug